Getting impress.js to load
- Can't put the script load in the body, because body is templated and
  therefore script tags are not evaluated.
- Can't put the script load in the head tag, because impress.js depends
  on document.body, which doesn't exist yet (meteor hasn't done magic
  yet).
- Can't use a document.write, element.append, or
  Template.rendered/created trick without doing something too hacky for
  my taste to determine when the script has loaded and is ready to have
  impress().init() called.

So I am wrapping impress.js with a Meteor.startup call, which should
cause the least amount of git merge conflicts in the future. It is late
enough for the body to exist... and by naming the file later
alphabetically than impress.js, the init call is consistently after
impress loads.

Moving the other script to the file as well.